% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_local.R
\name{source_local}
\alias{source_local}
\title{Source a file, temporarily using its directory as the local directory}
\usage{
source_local(file, ...)
}
\description{
Source a file, temporarily using its directory as the local directory
}
